Reworking workflows for Apache_local.hpf
This is largely a rework of the workflows which automatically create some of the Iris configurations, and also changes the Apache.hpf into Apache_Local.hpf.
Apache_Local has had Monitors 2 and 3 swapped so that it has a chance of working.  The installer now marks this profile for Experts Only
https://github.com/BlueFinBima/DCS-AH-64D-Profile/issues/9